refactor vtui
